Add ads to town generator
Yeah - to be able to afford the server that runs the website, I've decided to add a few ads to the website.
But don't worry, they're turned off by default so you won't be bothered by them locally. Plus, they don't show up on the town pages, only on the index page and the contributors list.